requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.9.5,<2.10
drf-spectacular>=0.25.1,<0.26
Pillow>=9.3.0,<10.0.0